Added quantlib tools package, requirements
Added quantlib_tools package and modified requirements for support.

Full testing is not yet configured, but was able to successfully pass
test cases present in twcli/core/quantlib_tools/test_models.py.

An error was raised when trying to build the environment for testing,
the package pytest and a couple of others were listed in both
requirements.txt and requirements-dev.txt. The duplicated packages were
removed, which suppressed the error.